+++ modules/misc/Portal.php
@@ -1064,9 +1064,9 @@ function PHPCheck()
 {
 	$ret = [];
 
-	if ( version_compare( PHP_VERSION, '5.4.45' ) == -1 )
+	if ( version_compare( PHP_VERSION, '5.5.9' ) == -1 )
 	{
-		$ret[] = 'RosarioSIS requires PHP 5.4.45 to run, your version is : ' . PHP_VERSION;
+		$ret[] = 'RosarioSIS requires PHP 5.5.9 to run, your version is : ' . PHP_VERSION;
 	}
 
 	if ( (bool) ini_get( 'safe_mode' ) )