+++ modules/Student_Billing/functions.inc.php
@@ -358,6 +358,12 @@ function _makeFeesFileInput( $value, $column )
 		);
 	}
 
+	if ( ! empty( $_REQUEST['LO_save'] ) )
+	{
+		// Export list.
+		return $value;
+	}
+
 	$file_path = $value;
 
 	$file_name = mb_substr( mb_strrchr( $file_path, '/' ), 1 );